+++ clang/docs/ReleaseNotes.rst
@@ -122,6 +122,19 @@ C23 Feature Support
 - No longer diagnose use of binary literals as an extension in C23 mode. Fixes
   `#72017 <https://github.com/llvm/llvm-project/issues/72017>`_.
 
+- Corrected parsing behavior for the ``alignas`` specifier/qualifier in C23. We
+  previously handled it as an attribute as in C++, but there are parsing
+  differences. The behavioral differences are:
+
+  .. code-block:: c
+
+     struct alignas(8) /* was accepted, now rejected */ S {
+       char alignas(8) /* was rejected, now accepted */ C;
+     };
+     int i alignas(8) /* was accepted, now rejected */ ;
+
+  Fixes (`#81472 <https://github.com/llvm/llvm-project/issues/81472>`_).
+
 Non-comprehensive list of changes in this release
 -------------------------------------------------